+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2025-02-22
+### Fixed
+- fixed interfaces to be compatible with Fable JS & TS
+### Added
+- TryPop function
+## Changed
+- in IDictionary rename Get ans Set to GetValue and SetValue
+
 ## [0.2.1] - 2024-10-30
 ### Changed
 - Removed IDictionary and IEnumerable interface becaus not compatible with Fable JS & TS yet
@@ -22,8 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added more tests
 
 
-[Unreleased]: https://github.com/goswinr/Dicts/compare/0.2.1...HEAD
-[0.2.0]: https://github.com/goswinr/Dicts/compare/0.1.0...0.2.1
+[Unreleased]: https://github.com/goswinr/Dicts/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/goswinr/Dicts/compare/0.2.1...0.3.0
+[0.2.1]: https://github.com/goswinr/Dicts/compare/0.1.0...0.2.1
 [0.1.0]: https://github.com/goswinr/Dicts/releases/tag/0.1.0
 
 <!--